+++ src/main/java/com/simplythewest/utd/Main.java
@@ -1,0 +1,4 @@
+package com.simplythewest.utd;
+
+public class Main {
+}
