--- src/main/java/ultimate_to_do/Main.java
@@ -1,4 +1,0 @@
-package ultimate_to_do;
-
-public class Main {
-}
